+++ Assets/LDtkUnity/Editor/Builders/TileBuildingJob.cs
@@ -1,4 +1,5 @@
-﻿using Unity.Collections;
+﻿using System;
+using Unity.Collections;
 using Unity.Jobs;
 using UnityEngine;
 
@@ -11,8 +12,7 @@ namespace LDtkUnity.Editor
             public int CoordId;
             public int PxX;
             public int PxY;
-            public bool FlipX;
-            public bool FlipY;
+            public int Flip;
         }
 
         public struct OutputData
@@ -49,8 +49,7 @@ namespace LDtkUnity.Editor
                     CoordId = isAutoLayer ? tile.D[1] : tile.D[0],
                     PxX = tile.Px[0],
                     PxY = tile.Px[1],
-                    FlipX = tile.FlipX,
-                    FlipY = tile.FlipY,
+                    Flip = tile.F,
                 };
             }
         }
@@ -69,9 +68,22 @@ namespace LDtkUnity.Editor
             offset.x = pxOffsetX / (float)LayerGridSize;
             offset.y = -pxOffsetY / (float)LayerGridSize;
             
+            //Rules can have multiple tiles built (like a 2x2 of art), but they all occupy the same coordId despite being located full cell(s) away!
+            //this results in offsets that can exceed 1 or -1, which at that point, should occupy the next cell over.
+            //not only is it easier to track down the tile in the editor, but it renders in a better z order with other tiles in that other cell.
+            int cellShiftX = offset.x > 0 ? (int)Math.Floor(offset.x) : (int)Math.Ceiling(offset.x);
+            int cellShiftY = offset.y > 0 ? (int)Math.Floor(offset.y) : (int)Math.Ceiling(offset.y);
+            cX += cellShiftX;
+            cY -= cellShiftY;
+            offset.x -= cellShiftX;
+            offset.y -= cellShiftY;
+            
+            bool flipX = (input.Flip & 1) == 1;
+            bool flipY = (input.Flip & 2) == 2;
+            
             Vector3 scale = new Vector3(ScaleFactor, ScaleFactor, 1);
-            scale.x *= input.FlipX ? -1 : 1;
-            scale.y *= input.FlipY ? -1 : 1;
+            scale.x *= flipX ? -1 : 1;
+            scale.y *= flipY ? -1 : 1;
             
             //convert y into unity tilemap coordinate space
             cY = -cY + LayerCHei - 1;